master: Finish refactoring of matocsserv.h
In one of the the previous commits a forward declaration of struct
matocsserventry was placed in matocsserv.h. This introduced an
inconsistency in a code -- the same objects were used in chunks.cc as
void* and matocsserv* pointers. This commit finishes refactoring of
matocsserv.h header by changing all the functions in this file to
operate on matocsserv* pointers. The header is also being reformatted.
The rest of the changes is needed in order to make the code compile
after the change in matocsserv.h (all uses of void* pointing to
matocsserv entry are replaced with matocsserventry*).

Change-Id: Idc8ba5416514298cdb7caf4189ca347d0d72ebd6