master: Modify chunks' replication algorithm to respect labels
This commit modifies the chunk loop in the master server to properly
handle labels when replicating missing copies and removing superfluous
copies. This enables almost a full support for custom goals in master
server - it will maintain the requested number of copies and will
properly handle changes in goal definitions or changes in files' goals.
The only thing that still might not work as expected is rebalancing of
chunks from servers with high disk usage to servers with low disk usage.

Change-Id: I1443f861358818e17beeb90fa4ea367ad57e5430