master: Modify algorithm of creating new chunks to respect labels
This commit changes the algorithm which chooses chunkservers to place
a new chunk. Now it makes use of custom goals and media labels.

Te algorithm is changed in a way that it no longer maintains 'carry'
between runs, which means that for each new chunk each chunkserver is
chosen completely randomly (with respect to weights equal its total disk
space though). The previous implementation was trying to distribute
chunks more evenly, but making it random shouldn't influence the
observed behavior of the system.

Change-Id: I335e6b116fa9c1678665781cf52f675533c316d0