master: Clean up code which counts copies of chunks
This commit refactors code which touches the matrix of chunks and
counters of chunk copies to remove literal constants like '10' or '11'
from the code. This is to help people distinguish between '10'
understood as "maximum possible goal" (which is no longer true) and '10'
understood as "maximum number of copies of a chunk that we maintain
counters for". To achieve this there is a new CHUNK_MATRIX_SIZE constant
defined in MFSCommunication.h file.

This commit also adds parsable descriptions (in MFSCommunication.h) for
network messages concerning the matrix of chunks and number of copies of
chunks for a given file.

Change-Id: I5e064b04a3ea6ea2974038eb2494c15ea0190bcf