master: Fix handling CLTOMA_CSSERV_REMOVESERV packets
Due to a bug in the master server's code removing servers from the list
in the CGI interface didn't work (and could even result in segmenattion
fault). This commit fixes the issue and also fixes descriptions of
CLTOMA_CSSERV_REMOVESERV and MATOCL_CSSERV_REMOVESERV network messages
in the MFSCommunication.h file.

Change-Id: I6cd8bddf7a2b091797437d77ccf601940ddea2d6