probe: Fix handling disconnected chunkservers
This commit fixes two commands: list-chunkservers and list-disks. If
a chunkserver disconnects, it is still present in the master server's
database. Such a server is reported in MATOCL_CSERV_LIST as a server
with a version equal to 256.0.0, with 0 chunks, 0 space, 0 errors, etc.
This commit fixes:
* the list-disks tool not to try to connect with disconnected servers to
  get a list of disks from them;
* the list-chunkservers command not to print this magic version string
  ("-" is printed instead) or zeros in other columns in the
  human-readable mode ("-" is printed instead)

This breaks a bit the compatibility of the --porcelain format, but
hopefully it's not too late to do this.

Change-Id: I315d0b0a77a04c23f396eedcff1591237c7248ff