master: Implement rebalancing of chunks which respects labels
This commit implements a new algorithm of chunk rebalancing which works
fine when custom goals are used. The new semantics of the
ACCEPTABLE_DIFFERENCE entry in the master server's configuration file is
as follows: we allow moving chunks between chunkservers to make chunk
distribution more even only if the source server has disk usage greater
then the destination server at least by ACCEPTABLE_DIFFERENCE (and, of
course, restrictions set by the goal are still met).

A new configuration option is introduced: if
CHUNKS_REBALANCING_BETWEEN_LABELS is set to 0 (which is the default), no
chunks will be moved between servers with different labels.

Change-Id: I8288b48294ad15c7c51aaae296a2e34aa865b192