mount: Fix printing session limitations after mounting
This commit fixes a message that is being printed after LizardFS is
successfully mounted. The message has a form of:

  mfsmaster accepted connection with parameters: read-write,restricted_ip,map_all ; root mapped to root:root ; users mapped to user:group ; setgoal limited to (2:8)

Due to an error in parsing a message from mfsmaster, the informations
about mapped users, goal or trashtime restrictions were incorrect. This
commit fixes the issue, so the mechanism should work OK now.

This commit also adds a dissection information for the
MATOCL_FUSE_REGISTER message (the one that was faultily parsed by
mount) to the Wireshark plugin.

Change-Id: I22dfabd5d8ba9bbf348e9c3c1d86cfa05de787bf