probe,master: Add chunks-health command to probe
Add command returning chunks health report to lizardfs-probe tool.
The new command can return 3 types of reports.

First, for every goal it can report how many chunks in this goal are
safe/endangered/lost.

Second, for every goal it can report how many chunks in this goal should
be replicated, taking into account the new custom goals feature.

Third, for every goal it can report how many chunks in this goal should be
deleted, also taking into account the new custom goals.

E.g. if a chunk is supposed to be stored in one copy on chunkserver
labeled 'hdd' and in another copy on chunkserver labeled 'ssd' but is
currently being stored on 2 chunkservers without any labels, the tool
will report that this chunk is safe, should be replicated twice (to
servers labeled 'hdd' and 'ssd') and also that two existing copies of
this chunk should be deleted (the ones stored on chunkservers without
any labels, which should be deleted after replicating to proper
servers).

Change-Id: Ic622e9371dfecadb1e98d65d4d59b0b51210b446