cgi: Add vim modeline to mfs.cgi.in file
This file is formatted against python standards defined in PEP8
(http://legacy.python.org/dev/peps/pep-0008/#indentation) and this makes
it difficult to edit with vim, which follows these standards -- the
default ftplugin/python.vim file contains:

  " As suggested by PEP8.
  setlocal expandtab shiftwidth=4 softtabstop=4 tabstop=8

This modeline overrides vim defaults to match the actual formatting of
the file.

This commit also fixes the formatting in this file where it was broken
by vim users.

Change-Id: I23adc6030c447b2ad8c8dda72cc4572fb90c0d2a