cgi: Add deserialization framework
This commit adds two function helpful when handling network messages:
* make_liz_message - to serialize a network message
* deserialize - to deserialize a response

Implementation of deserialize allows users to do something like:

  ret = deserialize(buffer, List(Tuple("LLL") + List(String)))

A network message LIZ_CLTOMA_LIST_GOALS is rewritten to use this
framework (because it would be reused in the near future to implement
other tabs in the web interface)

This commit also fixes a bug in cgiserv.py -- it adds a 'globals'
parameter equal to an empty dictionary to execfile() in cgiserv.py. It
makes mfs.cgi script run in it's own environment (not in environment of
a function which calls execfile) which makes it possible to use global
names normally in mfs.cgi.

Change-Id: I122f13229ca75f8bed8bed7e54d23c1d30d37eed