cgi: Add "Chunks" tabs
This commit adds a tab with information about availability of chunks for
each goal: how many chunks are safe, endangered or lost, how many copies
are will be replicated and how many copies will be removed. This makes
it possible to track state of the system which uses custom goals -- the
chunk matrix from "Info" tab doesn't present all the needed information.

Change-Id: I3d86f2d0e6254d8cad944cbcab4d3b23f5e156e3