common: Simplify format of the serialized GoalMap
This commit modifies the format of a serialized GoalMap. The previos one
was something like:

  pair<map<uint8_t, T>, T>

where the second T is a value kept for goal 0. This commit changes it
to:

  map<uint8_t, T>

The value for goal 0 is kept in the map like all other values. This make
the serialized maps simpler to parse, e.g. in the CGI interface.

The code using the old format hasn't been released yet, so this change
doesn't break any backward compatibility.

Change-Id: I700a058ca97efab1005807e0da8c0b3b114cca41