cgi: Fix restarting CGI server
In some cases restarting CGI server failed with the following error:

  sending SIGTERM to lock owner (pid:79276)
  Traceback (most recent call last):
    File "/usr/sbin/mfscgiserv", line 492, in <module>
      if wdlock(lockfname,mode,locktimeout)==1:
    File "/usr/sbin/mfscgiserv", line 419, in wdlock
      posix.kill(l,signal.SIGTERM)
  OSError: [Errno 3] No such process

This is because cgiserver tries to overwrite contents of the lockfile
with its own pid after killing the owner and in the following example:

  old pid: 12345
  new pid: 9876

the file was left with "98765" written inside. This causes

  mfscgiserver restart

to send TERM to a non-existing process and to fail. This commit fixes
the issue by adding 10 spaces after the pid to erase any digits.

Change-Id: If394b1c7ad028741ce7e4c5d8f39b02961d97bd8