debian: Fix lizardfs-master.init
This commit removes possibility for lizardfs-master.init to create a
new metadata.mfs file for the master server. This is to avoid problems
when master server's data path is located on a separate disk or
partition which failed to mount at boot time. lizardfs-master service
mustn't to start with empty metadata in such a scenario.


Change-Id: Id9f917df33a404ca24cecdf1c5f0129549c2765a