tests: Fix failing tests which rename files in trash
Tests which rename files in trash and immediately try to undelete them
might fail because of a race. The following code:

  echo new_name > /mnt/lzfs-meta/trash/*some_file
  mv /mnt/lzfs-meta/trash/*new_name /mnt/lzfs-meta/trash/undel

has a race: the file in trash is renamed in the fuse_release handler,
which is called just after close() on a file written by echo returns.
This makes it run in parallel with the forthcoming mv command and
(especially when using valgrind) causes mv to fail if fuse_rename is
executed before fuse_release finishes.

This commit fixes races in tests which rename files in trash by adding
assert_eventually (which waits for rename to finish) after each rename
operation on files from trash.

A new test is attached to this commit. This tests uses the rename
feature in LizardFS trash a lot to verify if it works properly. The race
can be observed if the 'assert_eventually "test -e (...)"' is commented
out and the test is executed with USE_VALGRIND=YES.

Change-Id: I2aed03310cc0e31fd67a45f711b52a5924983935