probe, tools: Fix handling ANTOAN_NOP packets
lizardfs-probe and mfstools used to fail if received ANTOAN_NOP instead
of the requested response. This commit fixes the issue by modyfing the
default behavior of the ServerConnection class to ignore ANTOAN_NOP
responses.

This commit also adds some code which can be used to create unit
tests of classes which use network connections (like ServerConnection)
and uses this code to test the ServerConnection class.

Change-Id: I8825350dd323a1efea92c88416366d6b63daa1b0