tests: Move test_acl_permissions.sh to LongSystemTests
This test cannot be run a machine configured only using
setup_machine.sh, because it requires a special version of the Linux
kernel. This makes it very inconvenient to run ShortSystemTests -- at
least this one test fails in most cases. This commit moves it to
LongSystemTests to avoid this problem.

Change-Id: I39f8473de1ef5c5c62ec8a2d2f657feb5cdb52a6