metarestore: Fix uncaught exceptions when applying changelogs
When metarestore is applying changelog files, uncaught exception can
occur when getting version of the last change. This would cause metarestore
to terminate in some cases eg. when one of the changelog files is malformed.

This commit fixes the bug (introduced in e701e69) by catching the
exception and setting the version of the last change to 0.

Change-Id: I28fe9f5d87d42ba34b4b94e61e15d87b4f19a730