tests: Fix failing test_shadow_reloading_metadata
After a change introduced in b75dac65879304923 this test fails because
of a timeout: master server refuses to terminate quickly if any
connected shadow master is not responding. This commit fixes the issue
by changing a master<->shadow timeout from the default value (60
seconds) to a minimum allowed value (10 seconds). After this change
stopping the master server takes 10 seconds and the test succeeds.

Change-Id: I32c63ab0dcf8f6a2ba69cfa41ab5651fd805fb82