deb, rpm, cmake: Fix building packages
This commit fixes the following issues with packages:
* RPM packages failed to build because of unpackaged manapages -- this
  commit adds them to appropriate packages
* RPM packages failed to build because of unpackaged ha-cluster files --
  this commit adds a possibility to disable processing the ha-cluster
  directory and disables it when generating RPMs, because the ha-cluster
  feature is not yet available on .rpm platforms
* .deb packages failed to build because of wrong filenames in the
  lizardfs-cluster-agents.install file -- this commit fixes it
* Generating manapages produced a very verbose output which isn't
  necessary when building packages, this commit makes the verbose output
  disabled by default (it useful only when writing manpages)
* Generating manpages failed on CentOS 6 because of unsupported option
  a2x --keep-artifacts -- this commit fixes this by disabling this
  option if verbose output is disabled (and it is disabled when building
  packages), because these artifacts may only be useful when writing
  manpages.
* Compilation of slogger.cc generated a warning:
    std::string syslogLevelToString(int)' defined but not used
  when compiling without DEBUG_LOG (e.g., when building packages), so
  the function which uses syslogLevelToString was reorganized to avoid
  the warning.

After introducing a possibility do disable ha-cluster in the 'configure'
script, also generating documentation can be disabled. This is useful
for people who want to build LizardFS using the configure script, but
don't want to download hundreds of megabytes of LaTeX packages required
by asciidoc.

This commit modified the 'configure' wrapper script not to ignore its
parameters, so they are verified now and the 'configure' script fails if
any unrecognized option is present. This is to warn people who think
that things like --prefix= may work. Because of this fact default calls
to this script were overridden in both 'rules' and 'lizardfs.spec'
files.

Change-Id: I6d29e6d27015b0089926e2a44831931fd110e510