ha-cluster: Make the cluster manager more user friendly
This commit modifies lizardfs-cluster-manager to print information about
progress of its work. It also modifies it to make --dry-run work as
expected, e.g.:
* not to call any mktemp which otherwise fail
* not to call any other commands (like grep on non-existent files)
  without invoke
* not to use any redirects like invode sed ... > file (these create
  empty files)

This commit also modifies output or --dry-run and --verbose to add
a prefix to each message printed by this script to be able to easily
distinguish between what the script prints and what commands executed by
it print.

Change-Id: I19068180a9a051134ab9a63911a2dfc182b52f80