master: Fix state reported by LIZ_MATOCL_METADATASERVER_STATUS
This commit fixes state reported by shadow master servers which can be
obtained using

  lizardfs-admin metadataserver-status

Shadow master servers used to report 'shadow connected' always when
a file descriptor of the socket used for a connection with master server
is opened. This was true just after sending SYN packet. This commit
modifies this behavior to report 'connected' only when a connection is
working and shadow master managed to register properly.

Change-Id: I90978c9e4da10cee915480d2534365ad71984ff7