master, admin: Add 'lizardfs-admin relaod-config' command
This commit adds a new 'reload-config' command in lizardfs-admin. This
command can be used to request reloading of the config file in the
shadow master (works for any personality). The command is synchronous,
i.e. it waits for the requested reload to complete.

A simple test of the new command and updated man pages for
lizardfs-admin are also provided by this commit.

Change-Id: I13b187c172921f189203ef85d31551a23f428382