admin: Refactor code used to authenticate admins
To make life easier, this commit refactors code of commands which use
authenticated connections in lizardfs-admin: a new simple class
RegisteredAdminConnection is introduced. This refactoring revealed a bug
in promote-shadow command - a response to LIZ_CLTOMA_ADMIN_BECOME_MASTER
was ignored and status returned by the challenge-response mechanism was
checked twice instead.

Change-Id: Id69c64bcd902381cdd9296cac9fdc068d5fe32d5