admin: Modify lizardfs-admin to ask for a password later
This commit modifies commands which use admin's password to ask for the
password after the connection is successfully established. To achieve
this, ANTOAN_NOP messages need to be send when the password is being
entered by a user to avoid a timeout. This feature (sending NOPs) will
also be useful in the future to implement commands which take a lot of
time on the master server's side to execute before sending a reply.

Change-Id: I211559b8ed206ceb46ffebcadd414ceb67ec0f59