common: Fix handling NOP messages in ServerConnection
This commit fixes an issue with ServerConnection::sendAndReceive called
with receiveMode=ReceiveMode::kReceiveFirstNonNopMessage. When it
received many NOP messages in a single read(sock,...) call, it did throw
anyway because it ignored only one NOP message. This commit fixes the
issue and now ServerConnection properly ignores any number of NOP
messages. A unit test is also modified to test such a behavior.

Change-Id: Ia833d712c4ba9c2abc296f89c38a3ce56a07578c