master: Change error code in LIZ_CLTOMA_ADMIN_REGISTER_RESPONSE
This commit changes the error code returned by a metadata server when
admin tries to register using a wrong password. The code is changed from
ERROR_EPERM to ERROR_BADPASSWORD. This makes the error message more
clear and makes it possible to distinguish from a situation when admin
access is disabled (i.e., no password is set).

Change-Id: I1e1d43fd39220652206cdf6b4ebab91ad635138e