chunkserver: Increase default for HDD_LEAVE_SPACE_DEFAULT to 4GiB
This commit changes the default value for HDD_LEAVE_SPACE_DEFAULT from
256MiB to 4GiB. This change was suggested by users who encounter
performance issues when disk usage on some chunkserver was close too
100%. These issues are caused by file fragmentation and other
filesystem specific problems. This can be avoided by not choosing
disks which are almost full for new chunks and the
HDD_LEAVE_SPACE_DEFAULT is meant to address this in one of possible
ways. The chosen value (4GiB) should be OK in much more cases than the
old value (256MiB).

Change-Id: Id2e71f4a70ccba376f9c88eb412d93efa5bf2300
Closes: #217