cgi: Replace mfscgiserv with lizardfs-cgiserver
This commit replaces the old mfscgiserv script and adds a new script
called lizardfs-cgiserver. The new script is almost a copy of the old
one, but some parts of code are removed. This is because the old script
includes a buggy code for service management, which can't be simply
removed without breaking backward compatibility. One of important bugs
is that such a server doesn't really require write access to any
directory, but the mfscgiserv needs it for creating its PID file. This
makes it impossible to run it as user 'nobody'.

The new script is a copy of the previous one, but:
* all the PID file management code is removed
* it doesn't have a daemon-like interface (start, stop, restart, ...);
  instead it just runs in foreground
* two small issues with the code are fixed to make it pass a test which
  uses pylint

This commit changes the implementation of init script for this service
on Debian and Ubuntu to the one prepared by Dmitry Smirnov
<onlyjob@member.fsf.org>.

The lizardfs-cgiserv.service file is also modified to use the new script
instead of the old one.

SysV scripts for CentOS/RHEL are temporarily left unchanged.

Tests are modified so that if a test uses CGI_SERVER=yes parameter for
setup_local_empty_lizardfs, the new implementation will be used.

The old mfscgiserv script should be marked as deprecated as soon as it
is no longer used by CentOS/RHEL SysV scripts.

Closes #214
Closes #215
Closes #219

Change-Id: I5df243867208526a187713c8b3ba272adead6b79