common: Fix compatibility with different libc implementations
This commit fixes an issue with the serialization framework which
assumed that int8_t, uint8_t and char are defined as three different
types. This is not true e.g. on some Solaris systems. This commit fixes
the issue by changing the assumption -- now the framerwork assumes that
char, unsigned char and signed char are three different types (which is
guaranteed by C++ standard) and that uint8_t and int8_t are defined as
one of those three (which is not guaranteed, but should be true on most
architectures).

Change-Id: I73de18af2ac1a7e0a9790a41605fa6a139eda511