common: Use fcntl instead of flock in Lockfile::lock
This commit replaces the implementation of Lockfile::lock to the one
which uses fcntl(..., F_SETLK, ...). The new implementation is more
portable and should work not only on Linux, but also on Solaris.

This commit includes also a test of metadata file locks, because this
mechanism wasn't covered by any test.

Change-Id: I19e3652ffa2912e0978a02b02297db4408cbdd5e