cgi: Add daemonization to lizardfs-cgiserver
This commit allows lizardfs-cgiserver to put itself in background and create
a pidfile. Daemonization occurs only if pidfile path is passed as argument,
which will be used only in CentOS6.4 init script. This solution does not
interfere with existing init scripts, because when "-p" option is not passed,
lizardfs-cgiserver's behaviour remains unchanged.

Change-Id: I6239dc12a15c20f9f3a1c55d3913cdd1757d95de