common: Use std::pow/log instead of pow/log for ints
This commit fixes an issue which cause LizardFS not to compile on
certain platforms because of errors like this:

  human_readable_format.cc:21:55: error: call of overloaded 'log(const uint64_t&)' is ambiguous
  human_readable_format.cc:21:67: error: call of overloaded 'log(const uint16_t&)' is ambiguous
  human_readable_format.cc:22:46: error: call of overloaded 'pow(const uint16_t&, uint16_t&)' is ambiguous

This commit fixes the issue by switching from 'pow' to 'std::pow' and
from 'log' to 'std::log'. C++11 introduced a special overload of these
two functions for integral types in the 'std' namespace and this
overload should be better than pow/log functions from the global
namespace.

Change-Id: I4b3d63ffce7c1d689152a364e1de746275fef931