tests: Fix race in tests using lizardfs-cgiserver
This commit fixes tests which use the CGI server (currently only one
test: test_cgi_validate_html). These tests could fail if wget is called
before the CGI server opens a socket. This might happen, because the
server was started using:

  lizardfs-cgiserver &

This commit fixes the race by using the new feature of the
lizardfs-cgiserver which allows it to daemonize after opening the
socket.

Change-Id: I9071e65da040860cf77813a2ebdfd9bffe5b5845