doc: Remove information about LOCK_FILE config option
This commit removes information about the LOCK_FILE config option from
master's, chunkserver's and metalogger's manuals as well as from their
config templates. This is because this option is not only deprecated,
but also removed a long time ago (in 412d9ed6d2f7d248).

RUN_PATH and RUN_SUBDIR variables from the build system and packages are
also removed, because they are no longer used anywhere and there are no
new use cases for them.

Change-Id: I43f053a0c86115aa21469efa28d3ac724b8c9455