chunkserver: Fix handling pread/pwrite errors
This commit fixes the way chunkserver tests if pread/pwrite failed. It
should do this just after calling these functions to use the proper
value of errno. Functions like gettimeofday or pthread_mutex_lock are
allowed to change errno, so we should avoiding overwriting errno by
them. Additionally when calling pread(), CRC of data should not be
calculated and tested if pread() failed.

Change-Id: I8f84ed4df0b22be9bad36923ef41a9498d4f6bcb