cgi: Add list of metadata servers to the Servers tab
This commit adds a list of all metadata servers to the Servers tab. In
the previous LizardFS release shadow masters were listed on this tab in
the 'Metadata Backup Loggers' table. In a87eba68b7cc5 shadow masters were
removed from this list and this commit restores them, so in the next
release information about shadow masters in the CGI interface should be
more comprehensive.

The new table in the CGI interface displays the same information as

  lizardfs-admin list-metadataservers

The command line tool was buggy (wrong host names were shown), so this
commit also fixes it.

Closes #243

Change-Id: I43d637fd4fbd3ebb8b04febd0f6af21abd304afe