cgi: Fix 'ERROR 403: Forbidden' when paths contain symlinks
This commit fixes an issue with CGI servers installed in paths which
contain symlinks. In such environments, the CGI server was responding
with 'HTTP/1.1 403 Forbidden' due to a mismatch in the following code:

  file_path = os.path.realpath(os.path.join(HTTP.root,*self.path.split('/')))
  if not file_name.startswith(HTTP.root+os.path.sep) and not file_name==HTTP.root:
      return self.err_resp(403,'Forbidden')

This is because the default value for HTTP.root is not processed with
os.path.realpath, but file_path is. This may result in a mismatch, e.g.
when HTTP.root contains symlinks. This commit modifies the code to
always use os.path.realpath before setting the HTTP.root variable.

Change-Id: I42598606bc0496fd6d37533e6ae0e353c1e793f8