chunkserver: Make posix_fadvise(POSIX_FADV_DONTNEED) optional
In one of the previous commits a mechanism of reducing cache pressure
generated by chunkservers was introduced. It makes use of
POSIX_FADV_DONTNEED which is called on every chunk before closing it.
This might be helpful if there are other services running together with
chunkserver, because their data will be cached longer. Cached
chunkserver's data in some use cases is not reused, because clients
(mounts) also cache files.

This commit makes this mechanism optional: it has to be enabled by
setting HDD_ADVISE_NO_CACHE=1 in chunkserver's config file. By default
dropping caches is disabled (it will require some testing to decide if
enabling it by default won't hurt users too much, especially those who
run chunkservers on dedicated machines and enjoy caching).

Only chunks which were read by the background job which looks for CRC
errors are always removed from cache, because this job reads more or
less random chunks and it it not very likely that these random chunks
will be reused.

Closes #212
Closes #225

Change-Id: I35db05b99e5ee24ddf6f6f4bb2380b5f76af0445