master: Add support for sending files to tapeserver
This commit implements master's behaviour when goal is changed to tape (@).
Files intended to be sent to tapeserver are stored in periodically cleared
queue. List of such files is provided to tapeserver via PUT_FILES message.
Current implementation works only for 1 tapeserver connected to master,
support for multiple tapeservers will be implemented in next iteration.

Change-Id: Ib91263587a5e196e81459f489227b6c6424a3e8b