master: Fix bug which causes over-replication of chunks
This commit fixes a bug which results in creating superfluous chunk
replicas when custom goals are used and some requirements specified in
such goals cannot be met, e.g.:
* a goal requires copies on servers labelled 'A B C', but there are no
 'A' servers
* a goal requires copies on servers labelled 'A A A', but there is only
  one 'A' server available.
In the above scenario chunks were replicated to all available servers.

This commit also adds a new test which verifies if the bug no longer
exists.

Closes #245

Change-Id: Ib7a7d080b89770f410c99937138740199f232317