tests: Fix unmounting fuse filesystems
Sometimes FUSE filesystems in /proc/mounts are indicated by
'fuse.xxxxxxxxxxx' in the third field of an entry. The previous version
of a code which looks for fuse filesystems assumed that this is always
equal to 'fuse'. This commit changes this assumption to make it work
properly, when 'fuse.something' is present.

Change-Id: Idf2e6f1a85ad517f3baedf5c8bc984742064f100