master: Prioritize replication of endangered chunks
This commit introduces a mechanism of prioritizing endangered chunks during
chunk maintenance. Depending on the config value ENDANGERED_CHUNKS_PRIORITY,
appropriate percentage of chunks intended for processing would be taken
from endangered chunks queue. A chunk is added to this queue when it's goal
is greater than 1, but number of it's valid copies drops to 1.

By default, ENDANGERED_CHUNKS_PRIORITY is set to 0, causing no overhead,
because endangered chunks will not be collected in any queue.

This commit comes with 3 tests, which check whether lizardfs operates without
errors if endangered prioritization is turned on.

Closes #244

Change-Id: I1729844c64033dd9743e4723a7c94e47860127ea