metarestore: New option -g
This commit adds new functionality to metarestore.
It can be used to determine latest metadata version
that can be restored from disk. If metarestore fails
(ex. metadata files are missing or corrupted) it prints 0,
otherwise it prints number of determined version.
Version should be adequate to the result of master -o auto-recovery.

Change-Id: If6e675a7efd4a6352e361402e550f99581fa6a82