tests: Fix ShortSystemTests.test_shadow_reconnect
This commit fixes an issue with shadow_reconnect test, which checks if files
in working directory remained unchanged. Unfortunately, stats.mfs file
is created precisely once per hour, so if the test is run on the border
of 2 hours (e.g. 19:59:50-20:00:04), it fails.
A way to fix this issue is to ignore stats.mfs file while performing a check,
since its existence depends on the time of day.

Change-Id: I090dbe866ed37ffb6c5f2908653ccec11cb40cbf