master: Fix support for REPLICATIONS_DELAY_DISCONNECT
If some chunkserver disconnects, all replications are paused
for REPLICATIONS_DELAY_DISCONNECT seconds. It's not a proper behaviour.

From this commit, disconnecting a chunkserver will not stop data migration
to servers with some other label.
Moreover, when some chunk has more missing copies than the number
of disconnected chunkservers, it still will be replicated.

This commit comes with 3 tests.

Change-Id: I60c2859284ec137ea6d561fc75dd70094cf96a07