common: Add memory efficient implementation of std::vector
This patch adds implementation of class compact_vector designed for
memory efficiency. Compact vector allocates memory only for data that
need to be stored. This is opposed to std::vector where additional memory
might be allocated to improve computational complexity.
This means that for compact_vector capacity() is always equal to size().
Also on 64 bit architectures both pointer to data and vector size
are kept in single 64 bit variable.

Change-Id: I05ec2af21980f9b07486782a7f830e62ddd5b41d