master: Improve handling multiple goals
This commit introduces new method of handling multiple goals for chunks.
A chunk can have more than 1 goal if it was snapshotted.

Previous implementation assumed that if multiple goals map to one chunk,
goal with the highest id is chosen. The new idea is to prepare a dynamic
goal in case of multiple definitions - union of all definitions.
Example:
 - file1 has goal {label1, label2}
 - file2 is a snapshot of file1
 - file2's goal is set to {label2, label3}
Now shared chunks of file1 and file2 have goal {label1, label2, label3}.
After changing goal of file1 to {_}, shared chunks will use {label2, label3}.

This commit comes with an amendment to mfsmakesnapshot test, which now
covers basic operation on goals as well.

Closes #298

Change-Id: Ibde4e4fc4363e3225e01cb50d276f1665213a0b9