common: Add BerkeleyDB back-end to hstring
This commits enables BerkeleyDB-based back-end for hstring.
Data is stored in DB_HEAP configuration. It means, that instead of specifying
both key and value, only value is passed to a database - which returns its
unique identifier. The identifier consists of 32bit page id and 16bit index,
which leaves 16bits to store hash for fast comparison.

In order to use BerkeleyDB back-end, libdb library >= 11.2.5.2 must be present.

Change-Id: I35ab458441168d86133bfea1b56360e4bb0eb921