cmake: Improve CheckCXXExpression macro
CheckCXXExpression was executing test program to determine
if specified expression was true. This causes problems with
cross compiling where it's not possible to run program
on destination machine. The macro is modified so it just
check if the test program can compile. Based on the compilation
result macro determines expression value.

Change-Id: Ibdc0d1312062197024e2bb087882a0b8a1b60502