master: Add implementation of POSIX locks
This commit introduces handling of POSIX advisory locks (Ref: man fcntl).
Interface allows to place a shared lock or an exclusive lock on a range
of target file. Overwriting owned locks is possible, as long as they
do not conflict with other locks placed on a file.

Lock's owner is identified by 64bit 'owner' field provided by FUSE,
session id and pid of a calling process.

Change-Id: I6f53ceda0ea29cc5ded391f5cd5516e54fad6def