master, tools: Remove support for +/- flags in setgoal
This commit removes support for increasing and
decreasing goal modifiers in setgoal.
With goal labels and different types of goals (like xor)
increasing/decreasing goal id leads to unpredictable
behaviour. Because there is no sensible modification to make
this option work it was removed altogether.
Test validating this functionality is also changed.

Change-Id: Ieb79bc5a76db9e7465c18f7a4557627ee8ce2419