tests: Improve test_metadata_checksum_recalculation
Test metadata_checksum_recalculation is nondeterministic and relies
on changing metadata while master server recalculates their checksum.
Changes should be made to metadata which are already included in a new
checksum and to those which aren't.

This commit increases number of files and directories in the
installation what improves chances of changing metadata in expected way.

Change-Id: I27970306efcc569c89e7fee08a04ebc9963648cb