common, master: Modify Goal::Slice to use flat storage
This commit modifies Slice to use flat storage. All data
are kept in single instance of small_vector. This is extremaly
memory efficient (usually we keep all data in one cache line),
but it's not too easy to use. To simplify access we use combination
of vector_range and flat_map to make interface more similar
to previous implementation of Slice.
Additionally there are some changes in ChunkCopiesCalculator
to efficiently use new Slice class.

Change-Id: I1bbd235794b521c2aa53d3d28beed772897b9a94