chunkserver: Fix possible segmentation fault in chunk scan
This commit fixes segmentation fault in chunk scanning code.

The problem occured when two file copies of chunk existed.
The first copy was invalid (i.e. io error on trying to remove file)
and second was the correct one. During the scan the first file was added
as chunk. Later the second chunk was parsed and hdd_chunk_get kept
returning NULL because the first chunk was invalid.
Then, hdd_chunk_add() function used this pointer without checking
if it is valid, which caused segmentation fault.

The first part of this fix is that the function hdd_chunk_get returns
pointer to new chunk in such case (if the CH_NEW_NONE flag wasn't set).
The second part is that hdd_add_chunk checks for validity of
returned pointer.

Change-Id: I348667bc9565aec938b97aa5bab20866ea4c45af