master: Try to keep parts of the same chunk on different servers
This commit changes logic of picking chunkservers for replication.
If it's possible to replicate to a server that does not have _any_
copy of the chunk, it should be done. Only if there is no other option,
should copy of the same chunk, but from different slice be placed
on one chunkserver.

Closes #331

Change-Id: Ic1f30b4e23d6ce23bf5dd2034637532effec7509